requirements.txt
Django>=4.0
djangorestframework
psycopg2-binary
pandas
celery
redis
openpyxl
django_extensions